requirements.txt
streamlit
pandas
plotly
pymongo
python-dotenv
google-generativeai